Fix crash related to custom death events & other small stuff

+++ DemoParser.cpp
@@ -52,7 +52,7 @@ DemoParser::~DemoParser()
 	gpParser = nullptr;
 
 	const size_t num_ents = m_Entities.size();
-	for( int i = 0; i < num_ents; ++i )
+	for( size_t i = 0; i < num_ents; ++i )
 	{
 		delete m_Entities[ i ];
 	}
